Move FFI part for strings into its own module
